refactor(desktop): split preset settings view
