+++ apps/desktop/src/features/library-workspace/session-presets/model/types.ts
@@ -1,0 +1,1 @@
+export type EditorMode = 'session' | 'time-step';
